# Conflicts:
#	src/main/java/towssome/server/service/ViewlikeService.java

+++ src/main/java/towssome/server/controller/ReviewController.java
@@ -17,7 +17,6 @@ import towssome.server.service.ReviewPostService;
 import towssome.server.service.ViewlikeService;
 
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 
 @RestController
@@ -35,19 +34,19 @@ public class ReviewController {
     public ResponseEntity<?> createReview(@RequestPart(value = "body") ReviewPostReq req,
                                           @RequestPart(value = "photos", required = false) List<MultipartFile> photos) throws IOException { // additional implementation needed for session
         log.info("reviewPostDTO = {}", req);
-        ReviewPost reviewPost = reviewPostService.createReview(req);
-        photoService.saveReviewPhoto(photos, reviewPost);
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        reviewPostService.createReview(req, photos, username);
         return new ResponseEntity<>("Review Create Complete", HttpStatus.OK);
     }
 
+    /** 특정리뷰글 조회 */
     @GetMapping("/{reviewId}")
     public ResponseEntity<ReviewPostRes> getReview(@PathVariable Long reviewId){ // get review by reviewId
         String name = SecurityContextHolder.getContext().getAuthentication().getName();
-
         ReviewPost review = reviewPostService.getReview(reviewId);
         log.info("review = {}",review.getId());
         List<PhotoInPost> photo = photoService.getPhotoS3Path(review);
-        ReviewPostRes reviewRes = null;
+        ReviewPostRes reviewRes;
 
         //비회원 조회
         if (name.equals("anonymousUser")) {
@@ -65,10 +64,7 @@ public class ReviewController {
         }else {
             //회원 조회
             Member member = memberService.getMember(name);
-            viewlikeService.viewProcess(new ViewLikeReq(
-                    reviewId,
-                    member.getId()
-            ));
+            viewlikeService.viewProcess(review, member);
             reviewRes = new ReviewPostRes(
                     review.getBody(),
                     review.getPrice(),
@@ -85,39 +81,33 @@ public class ReviewController {
         return new ResponseEntity<>(reviewRes, HttpStatus.OK);
     }
 
-
     @PostMapping("/update/{reviewId}")
     public ResponseEntity<?> updateReview(@PathVariable Long reviewId,
-                                          @RequestBody ReviewPostUpdateReq req,
-                                          @RequestPart List<MultipartFile> addPhotos) throws IOException { // update review by reviewId
-        List<Long> deletedPhotoIds = req.willDeletePhoto();
-        for (Long deletedPhotoId : deletedPhotoIds) {
-            photoService.deletePhoto(deletedPhotoId);
+                                          @RequestPart(value = "body") ReviewPostUpdateReq req,
+                                          @RequestPart(value = "photos") List<MultipartFile> addPhotos) throws IOException { // update review by reviewId
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        if(!reviewPostService.getReview(reviewId).getMember().getUsername().equals(username)){
+            return new ResponseEntity<>("You are not the author of this review", HttpStatus.FORBIDDEN);
         }
-
-        ReviewPost post = reviewPostService.getReview(reviewId);
-        photoService.saveReviewPhoto(addPhotos, post);
-        ReviewPostUpdateDto dto = new ReviewPostUpdateDto(
-                reviewId,
-                req.body(),
-                req.price()
-        );
-        reviewPostService.updateReview(dto);
-
-        return new ResponseEntity<>(HttpStatus.OK);
+        reviewPostService.updateReview(reviewId, req, addPhotos);
+        return new ResponseEntity<>("update complete", HttpStatus.OK);
     }
 
     @PostMapping("/delete/{reviewId}")
     public ResponseEntity<?> deleteReview(@PathVariable Long reviewId) { // delete review by reviewId
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        if(!reviewPostService.getReview(reviewId).getMember().getUsername().equals(username)){
+            return new ResponseEntity<>("You are not the author of this review", HttpStatus.FORBIDDEN);
+        }
         ReviewPost review = reviewPostService.getReview(reviewId);
         reviewPostService.deleteReview(review);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @GetMapping // EX) review?cursorId=10&size=10 -> id 10번 리뷰글 id보다 작은 10개의 리뷰글(id = 1~9)을 가져옴
-    public CursorResult<ReviewPost> getReviews(Long cursorId, Integer size) { // get all review(size 만큼의 리뷰글과 다음 리뷰글의 존재여부(boolean) 전달)
+    public CursorResult<ReviewPostRes> getReviews(Long cursorId, Integer size) { // get all review(size 만큼의 리뷰글과 다음 리뷰글의 존재여부(boolean) 전달)
         if(size == null) size = PAGE_SIZE;
-        return this.reviewPostService.getReviewPage(cursorId, PageRequest.of(0, size));
+        return reviewPostService.getReviewPage(cursorId, PageRequest.of(0, size));
     }
 
 }
